fix(merge-next): proxyconfig
